Proyecto/requirements.txt
flask_login
flask_sqlalchemy
flask_bcrypt
pysnmp
pysnmp-mibs
yagmail
pygal
cairosvg
netmiko
networkx
apscheduler
